cleaned up warning when compiled with JPRDEBUG
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@91 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/jprdebug.h
@@ -16,7 +16,7 @@ DBGDUMP(const char* mem, size_t len)
    char hexbuf[3];
    int wrap = 0;
    char ascbuf[wrapsize+1];
-   int ix;
+   size_t ix;
 
    DBGMSG("* DEBUGDUMP BEGIN *\n");
